Cope with portable text introduction.
Extracted components to own files before reusing for intro.

Theoretically this will allow gradual migration, but with no facility for
draft content and the feature still behind a flag I'm just going to go
for it later this evening.

See https://github.com/microbit-foundation/python-editor-next/issues/332
